fix(failover): strip the effort suffix off pool-fallback targets before catalog resolution
