+++ src/cc-template.ts
@@ -12,6 +12,7 @@ import {
   loadTemplate, promptVariantsOf, TemplateData, VARIANT_FAMILIES,
   PLATFORM_ONLY_TOOLS, INTERACTIVE_ONLY_TOOLS, CONFIG_SCOPED_TOOLS,
 } from './live-fingerprint.js';
+import { VALID_EFFORT_VALUES, parseEffortSuffix, type EffortValue } from './effort.js';
 
 // Re-exported so existing importers (scripts/capture-and-bake.mjs, the template
 // invariant tests) keep their import site. The definitions live in
@@ -1274,31 +1275,15 @@ export function resolveMaxTokens(flag: number | 'client' | undefined, clientBody
   return flag;
 }
 
-/** Valid values for the `--effort` flag. Mirrors CC's effort set (`low|medium|high|xhigh|max`) plus CC's `ultracode` mode and dario's pseudo-value `'client'` for passthrough. `'ultracode'` is CC's xhigh-plus-dynamic-workflow-orchestration mode (CC 2.1.154); the Messages API accepts only low|medium|high|xhigh|max, so dario normalizes ultracode → 'xhigh' on the wire (see normalizeEffortForWire). `'client'` passes through the client's own `output_config.effort` (falling back to `'xhigh'`). dario#87, `'max'` added in dario#190, `'ultracode'` added 2026-05-28. */
-export type EffortValue = 'low' | 'medium' | 'high' | 'xhigh' | 'ultracode' | 'max' | 'client';
-export const VALID_EFFORT_VALUES: ReadonlyArray<EffortValue> = ['low', 'medium', 'high', 'xhigh', 'ultracode', 'max', 'client'];
-
-/**
- * dario#419 — strip an optional effort suffix off a model name, so OpenAI-compat
- * clients that can't set `output_config.effort` (e.g. Cursor) can choose effort
- * by model name: `opus-4-8:high` (colon) or Cursor-style `claude-opus-4-8-high`
- * (hyphen). Only the wire-valid effort levels are recognized as a suffix — any
- * other trailing token is left as part of the model name, and a bare model that
- * IS an effort word (e.g. just "high") is left alone. Returns the model with the
- * suffix removed plus the parsed effort (undefined when none). Exported for tests.
- */
-const SUFFIX_EFFORTS: ReadonlyArray<EffortValue> = ['ultracode', 'medium', 'xhigh', 'high', 'low', 'max'];
-export function parseEffortSuffix(model: string): { model: string; effort?: EffortValue } {
-  for (const e of SUFFIX_EFFORTS) {
-    for (const sep of [':', '-']) {
-      const tag = sep + e;
-      if (model.length > tag.length && model.endsWith(tag)) {
-        return { model: model.slice(0, -tag.length), effort: e };
-      }
-    }
-  }
-  return { model };
-}
+// The effort vocabulary and the model-name effort suffix live in effort.ts so
+// the pool-fallback classifier can parse the SAME suffix the request path
+// parses without importing this module — claude-model.ts stays decidable over
+// a base set alone, and hand-copying the suffix list there is what would let
+// `claude:opus:high` drift from `opus:high` (dario#1156 review). Re-exported
+// here because proxy.ts, cli.ts and test/effort-flag.mjs import them from
+// cc-template.
+export { VALID_EFFORT_VALUES, parseEffortSuffix };
+export type { EffortValue };
 
 /**
  * Normalize an effort value to a wire-valid `output_config.effort`. The
